refactor: 코드 중복 제거 및 공통 유틸리티 추출
- ResponsePage와 OrganizerSlotSelectionPage에서 중복되던 timeSlot 생성 로직을 공통 유틸리티로 추출
- apps/web/src/utils/timeSlot.ts 생성
- generateTimeSlots()와 isBlockedSlot() 함수 export
- TimeSlotData 인터페이스 export
- import 제거로 불필요한 의존성 줄임

- gemini-code-assist 리뷰어 반영: 코드 중복 해결
