+++ apps/web/src/pages/OrganizerSlotSelectionPage.tsx
@@ -1,12 +1,8 @@
-import { useState, useMemo, useEffect } from 'react';
+import { useState, useEffect } from 'react';
+import { generateTimeSlots, isBlockedSlot, type TimeSlotData } from '../utils/timeSlot';
 import { useNavigate, useLocation } from 'react-router-dom';
 import { AppBar, Toolbar, IconButton, Typography, Box, Button, CircularProgress } from '@mui/material';
 import ArrowBackIcon from '@mui/icons-material/ArrowBack';
-
-interface TimeSlotData {
-  date: string;
-  slots: string[];
-}
 
 interface MeetingFormData {
   title: string;
@@ -39,54 +35,7 @@ export default function OrganizerSlotSelectionPage() {
   const startDateString = meetingData?.startDate;
   const endDateString = meetingData?.endDate;
 
-  const timeSlotData = useMemo<TimeSlotData[]>(() => {
-    if (!startDateString || !endDateString) return [];
-
-    const startDate = new Date(startDateString);
-    const endDate = new Date(endDateString);
-
-    if (Number.isNaN(startDate.getTime()) || Number.isNaN(endDate.getTime())) {
-      return [];
-    }
-
-    const dates: string[] = [];
-    const current = new Date(startDate);
-    current.setHours(0, 0, 0, 0);
-
-    const end = new Date(endDate);
-    end.setHours(0, 0, 0, 0);
-
-    while (current <= end) {
-      const dayOfWeek = current.getDay();
-      if (dayOfWeek !== 0 && dayOfWeek !== 6) {
-        dates.push(current.toISOString());
-      }
-      current.setDate(current.getDate() + 1);
-    }
-
-    return dates.map((date) => {
-      const slots: string[] = [];
-
-      for (let hour = 9; hour < 18; hour++) {
-        if (hour === 12) continue;
-
-        for (let minute = 0; minute < 60; minute += 30) {
-          const timeStr = `${String(hour).padStart(2, '0')}:${String(minute).padStart(2, '0')}`;
-          const slotDate = new Date(date);
-          slotDate.setHours(hour, minute, 0, 0);
-          slots.push(`${slotDate.toISOString()}|${timeStr}`);
-        }
-      }
-
-      return { date, slots };
-    });
-  }, [startDateString, endDateString]);
-
-  const isBlockedSlot = (date: string, time: string): boolean => {
-    const hour = parseInt(time.split(':')[0]);
-    const dayOfWeek = new Date(date).getDay();
-    return hour === 12 || dayOfWeek === 0 || dayOfWeek === 6;
-  };
+  const timeSlotData = generateTimeSlots(startDateString || '', endDateString || '');
 
   const toggleSlot = (slotIso: string) => {
     setSelectedSlots((prev) => {
