+++ work-history/2026-01-14-work-plan.md
@@ -1,0 +1,114 @@
+# 2026-01-14 작업 계획
+
+## 오늘 완료된 작업
+
+### ✅ 주최자 시간 선택 기능 구현
+- **PR**: #12 (https://github.com/Crong-Gabia/BookingTime/pull/12)
+- **브랜치**: `feature/organizer-time-selection`
+- **구현 내용**:
+  - 주최자가 회의 생성 시 자신의 가능한 시간 먼저 선택
+  - 참여자는 주최자가 선택한 시간만 응답 가능
+  - 주최자 불가 시간 orange dashed border로 표시 및 disabled 처리
+
+## 오늘 해야 할 작업 (TODO)
+
+### 1순위: 빌드/테스트 환경 안정화 ⚠️ 긴급
+
+#### 1.1 TypeScript LSP 데코레이터 에러 수정
+- **파일**: `apps/api/src/modules/meeting/meeting.service.ts`, `meeting.controller.ts`, `health.controller.ts`
+- **문제**: `Decorators are not valid here` 오류
+- **해결**: tsconfig 설정 확인 및 데코레이터 사용 방식 수정
+
+#### 1.2 테스트 파일 vi 네임스페이스 오류 수정
+- **파일**: `apps/api/src/modules/meeting/health.test.ts`, `meeting.test.ts`
+- **문제**: `Cannot find namespace 'vi'` 오류
+- **해결**: Vitest 타입 설정 추가 (`vitest/globals`)
+
+#### 1.3 DashboardPage 타입 에러 수정
+- **파일**: `apps/web/src/pages/DashboardPage.tsx`
+- **문제**: 파라미터 `any` 타입 사용, 컴포넌트 props 미스매치
+- **해결**: 타입 명시적 선언, props 타입 확인
+
+#### 1.4 HomePage 미사용 변수 제거
+- **파일**: `apps/web/src/pages/HomePage.tsx`
+- **문제**: `Meeting`, `meetingsLoading` 선언 후 미사용
+- **해결**: 미사용 변수 삭제
+
+#### 1.5 DB/도커 환경 세팅 확인
+- **상태**: ✅ 도커 컨테이너 시작됨 (postgres, redis)
+- **확인 필요**:
+  - DB 마이그레이션 실행
+  - 개발 서버 정상 시작 확인
+  - API 엔드포인트 테스트
+
+### 2순위: 기능 개선
+
+#### 2.1 External Adapters 실제 구현 (Issue #11)
+- **RoomAdapter**: 회의실 예약 시스템 API 연동
+- **HolidayAdapter**: 공휴일 API 연동
+- **HRAdapter**: 사내 HR/근태 시스템 API 연동
+
+#### 2.2 응답 기한 강제 종료 기능
+- 기획: 매일 특정 시간에 자동 종료
+- 현재: 주최자가 수동으로 [확정] 버튼 누를 때만 종료
+- 필요: 타이머 기능 또는 기한 설정 UI 추가
+
+#### 2.3 알림 발송 연동 (메신저 봇)
+- 현재: `sendReminder()` API 존재하지만 실제 발송 없음
+- 필요: Slack/Teams/KakaoWorkplace 등 연동
+
+### 3순위: 기획서 기준 MVP 완성
+
+#### 3.1 미응답자 처리 로직
+- 기획: 시스템에서 강제 제외하지 않음
+- 현재: 미응답자도 공통 슬롯에 포함됨
+- 필요: 미응답자를 고려한 공통 슬롯 계산 또는 "제외하고 진행" 기능
+
+#### 3.2 회의실 예약 시스템 실제 연동
+- 기획: RoomAPI를 통해 회의실 예약 + 캘린더 등록
+- 현재: 단순 체크만 수행
+- 필요: 실제 예약 API 호출 + 캘린더 등록
+
+### 4순위: Phase 2 고도화 (장기 계획)
+
+#### 4.1 지능형 추천 (AI)
+- 참석자들이 주로 선호하는 시간대 추천
+- 회의실 동선 고려한 추천
+
+#### 4.2 LLM 기반 자연어 생성
+- "다음 주 김팀장님이랑 면접 잡아줘" 입력하면 폼 자동 채우기
+
+#### 4.3 회의실 자동 배정
+- 참석 인원수와 회의실 정원 비교하여 최적 회의실 자동 선택
+
+#### 4.4 외부 캘린더 연동
+- Google/네이버 캘린더 API 연동
+- 사외 일정까지 고려한 필터링
+
+---
+
+## 작업 우선순위 요약
+
+1. **즉시 해결 (오늘 목표)**
+   - ✅ 도커/DB 환경 세팅
+   - ⏳ TypeScript LSP 에러 수정
+   - ⏳ 테스트 vi 네임스페이스 오류 수정
+   - ⏳ DashboardPage 타입 에러 수정
+   - ⏳ 빌드 성공 확인
+
+2. **1순위 (이번 주)**
+   - Issue #11: External Adapters 실제 구현
+   - 응답 기한 강제 종료 기능
+   - 알림 발송 연동
+
+3. **2순위 (다음 주 이후)**
+   - MVP 완성 (미응답자 처리, 회의실 예약 연동)
+   - Phase 2 고도화 기능
+
+---
+
+## 참고
+
+- 기획서: `work-history/product-spec/plan.md`
+- 개발 가이드: `guidelines/README.md`
+- GitHub Issues: https://github.com/Crong-Gabia/BookingTime/issues
